requirements.txt
psycopg2-binary==2.8.3
django==1.11.23
gunicorn==19.9.0
social-auth-core==3.2.0
social-auth-app-django==3.1.0
vk==2.0.2
Pillow==6.1.0
django-select2==6.3.1
django-ckeditor==5.7.1
python-slugify==3.0.3

bleach==2.0.0
chardet==3.0.4
django-bleach==0.3.0
django-bootstrap3==9.0.0
django-registration-redux==1.3
django-sendfile==0.3.11
django-user-agents==0.3.0
django-flat-responsive==2.0
django-sitetree==1.8.0
dealer==2.0.5
docxtpl==0.3.9
html5lib==0.999999999
olefile==0.44
pypng==0.0.18
PyQRCode==1.1
python3-memcached==1.51
requests==2.18.4
ua-parser==0.3.6
Unidecode==0.4.18
urllib3==1.22
user-agents==0.3.2
webencodings==0.5.1
Petrovich==1.0.0
pytz==2017.2
pandas==0.20.3
raven==6.2.1
django-nested-inline==0.3.7
